Refactor handshake message types into separate module
